requirements.txt
torch
torchvision
pennylane
numpy
pillow
scikit-learn
gradio